requirements.txt
python-dotenv==1.0.1
langchain==0.3.12
langchain-community==0.3.12
langchain-openai==0.2.14
langchain-huggingface==0.1.2
chromadb==0.5.23
openai==1.59.3
sentence-transformers==3.3.1
torch==2.5.1
streamlit==1.41.0
